variants_db_lamindb_usecase/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mutation-db"
version = "0.1.0"
description = "A toolkit for VCF mutation database management with biological annotations"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "lamindb>=0.60.0",
    "bionty>=0.30.0",
    "cyvcf2>=0.30.0",
    "pandas>=1.5.0",
    "numpy>=1.20.0",
    "requests>=2.28.0",
    "biopython>=1.80",
    "gseapy>=1.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mutation_db", "mutation_db.*"]
exclude = ["mutation_db_bionty", "mutation_db_bionty.*", "data", "tests", "examples", "docs"]
